+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [6.1.0] - 2024-5-10
 
-* Adds `unsubscribeLink` parameter to `sendEmail`
+* Adds `oneClickUnsubscribeURL` parameter to `sendEmail`
 
 ## [6.0.0] - 2023-12-27